+++ src/program/ui/AnalogInputsModel.h
@@ -1,0 +1,52 @@
+/*
+    Copyright 2015-2024 Clément Gallet <clement.gallet@ens-lyon.org>
+
+    This file is part of libTAS.
+
+    libTAS is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    libTAS is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with libTAS.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef LIBTAS_ANALOGINPUTSMODEL_H_INCLUDED
+#define LIBTAS_ANALOGINPUTSMODEL_H_INCLUDED
+
+#include <QtCore/QAbstractListModel>
+#include <QtCore/QMimeData>
+
+/* Forward declaration */
+struct Context;
+
+class AnalogInputsModel : public QAbstractListModel {
+    Q_OBJECT
+
+public:
+    AnalogInputsModel(Context* c, QObject *parent = Q_NULLPTR);
+
+private:
+    Context *context;
+
+    int rowCount(const QModelIndex &parent = QModelIndex()) const override;
+
+    QVariant headerData(int section, Qt::Orientation orientation, int role) const override;
+
+    QVariant data(const QModelIndex &index, int role = Qt::DisplayRole) const override;
+
+    Qt::ItemFlags flags(const QModelIndex &index) const override;
+    
+    Qt::DropActions supportedDragActions() const override;
+
+    QStringList mimeTypes() const override;
+    QMimeData *mimeData(const QModelIndexList &indexes) const override;
+};
+
+#endif